refactor: move type cercion validators to config.py
Moves `parse_bool_or_str` and `parse_int_or_str` functions into `config.py` to handle flexible configuration inputs (e.g., strings like "true" or "1", or integers) when setting up LLM infrastructure.

These validators are used in `nuextract.py` to ensure configuration fields are correctly typed, and corresponding unit tests are added to `test_config.py` to verify the coercion logic and warning behavior.
